Handle format errors
Unhandled errors in this code cause the stream processing to fail and no event to come back to our client handler in RoboCleaner.ts This results in timeouts in the calling code. This patch adds error handling.

Testing: Added a unit tests for this case to RoboCleaner and it now returns correctly.